fix: responsive hide amount and nav buttons
